Fix select steps streaming test
For reader to be able to open file before writer has issued any put
calls (under DataSizeBased aggregation scheme), we have to change the
approach so that at least the metadata files get created as soon as
the writer is constructed.  The rest of the code from InitTransports()
can then remain deferred until the first write call.