bp5writer: InitTransports() allow opening a file later
Break out the bits of InitTransports that generate the filename
and open the file, and put them into a new function that can be
called either with or without an mpi communicator.

This supports the rerouting aggregator, currently under development,
by allowing rerouted ranks to change their target subfile near the
end of the writing process, after opening a different file at the
beginning.